chore: move meditatedRecently to ../helper
