+++ src/app/routes/_message.router.js
@@ -1,5 +1,5 @@
 import Message from '../models/message.model.js';
-import meditatedRecently from './meditatedRecently.js';
+import meditatedRecently from '../helper/meditatedRecently.js';
 
 import moment from 'moment';
